requirements.txt
# Core
python-dotenv>=1.0.0
supabase>=2.0.0
redis>=5.0.0
httpx>=0.27.0

# Scrapers
yt-dlp>=2024.1.0
youtube-comment-downloader>=0.1.0
youtube-transcript-api>=0.6.0
telethon>=1.35.0
playwright>=1.40.0
playwright-stealth>=1.0.0
praw>=7.7.0
twikit>=2.0.0
feedparser>=6.0.0
trafilatura>=1.8.0

# Transcription
openai-whisper>=20231117

# NLP / ML
transformers>=4.36.0
torch>=2.1.0
sentence-transformers>=2.3.0
hdbscan>=0.8.33
numpy>=1.26.0
scikit-learn>=1.3.0
datasketch>=1.6.0
fasttext-wheel>=0.9.2

# LLM
anthropic>=0.39.0

# Task Queue
celery[redis]>=5.3.0

# Utilities
loguru>=0.7.0
